fido: change 'incognito' to 'off-the-record' in a few places
It's more accurate, since we generally don't care which kind of
off-the-record context we're dealing with. (The exception to this is
when instantiating platform authenticators, which may be available in
Incognito but not Guest mode e.g.).

Also change plumbing of the off-the-record bit in
AuthenticatorRequestDialogModel to flow with the TransportAvailability
struct.

Bug: 1157651
Change-Id: I31485c734547b74b7d0e46534e0da5cbe4f1a00f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2587503
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840444}