Propagate left button flag to viz to stop Delegated Ink Trail
Part 2 of ensuring that a delegated ink trail stops as quickly as
possible when a user lifts the left mouse button or pen tip, part 1
here[1].

This CL checks the new flag on the RenderFrameMetadata in the browser
process to determine if the left button is expected to be down or not.
It compares that information to if the button is down on the current
point or not, and if they don't match, it informs viz to reset
prediction. Resetting prediction when we know the trail is ending will
help the trail better match the user's expectation of where the ink
should show up.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/2579704

Bug: 1052145
Change-Id: If2ffcba7dbddf326263bb982fc8b002d5b162221
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2581001
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#840463}