Add field is_dom_match to ClientPhishingRequest
This CL adds a new boolean field is_dom_match to CSD pings. This makes
it easier to separate out the pings due to the DOM model from the ones
due to the visual model. In principle, this can be computed by comparing
the score in the ping to the model threshold, but this process is fairly
complex due to the potential for models to change on the server-side.

Change-Id: I8685f5795eec41dddb433fe8f6dec5a4f30bee0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2610951
Reviewed-by: Bettina Dea <bdea@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840467}