Use ExecutionContext's BrowserInterfaceBroker to bind GamepadHapticsManager.
We are in the process of attributing as many mojo interfaces as possible
to specific execution contexts, which would enable better control by the
browser over those contexts. As part of that effort, this CL uses the
execution context's BrowserInterfaceBroker to bind the
GamepadHapticsManager remote and moves the code registering the binders
accordingly.

Bug: 1126758
Change-Id: I6fb734720d56581e7af36a869affe573fc3c4451
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2596709
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Commit-Queue: Tal Pressman <talp@google.com>
Cr-Commit-Position: refs/heads/master@{#840490}