[PartitionAlloc] Better freelist corruption detection.
Some PartitionAlloc crashes are likely due to freelist
corruption. PartitionAlloc has two types of freelists: in the central
allocator, and in the thread cache. The central allocator one already
has a DCHECK() to verify that subsequent entries belong the same
superpage, which doesn't hold for the thread cache freelists.

This CL:
- Makes these crash lead to a NOINLINE function
- Add an interity check to all freelists.

This is not meant to be a security mitigation, but to prevent against
accidental issues.

Bug: 998048
Change-Id: I21aedfe2b6363069362514a8edd6cd5bdea1acfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2582218
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840587}