Gracefully shutdown if display server dies on desktop Linux
On Linux, the display server closes when the session is ending.  In
particular, for X11 connections, the server may be remote, so the
connection can go down at any time.  [1] added a crash on connection
IO errors where it would previously gracefully exit.  This is desirable
on ChromeOS to generate crash reports, but not for desktop Linux, so
this CL restores the old behavior on Linux.

[1] https://chromium.googlesource.com/chromium/src/+/59a4564762df2b3cb5a6037d5ce97e5ad626eb48

R=rjkroege
BUG=1160140

Change-Id: Ie5a5ce762e6b65f533411ce88caa51d3c27366a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2600045
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840661}