Speculative fix: iframe hypertext returns null hyperlink at index 0
An iframe can only have 1 hyperlink. If the iframe is not loaded yet,
then it should have 0.

When hyperlink 0 is retrieved, it should never be null. However, this
is occurring where the url is "empty". It may be occurring during
document destruction.

There exists code to update the parent hypertext when a
BrowserAccessibilityManager's contents are attached, but not on
destruction. It's unclear whether a single BrowserAccessibilityManager
can be destroyed without the entire tree of them being destroyed.
To be safe, update the parent when a BrowserAccessibilityManager
is destroyed.

Bug: 1157731
Change-Id: I96a6654d3aa2d6934c07a0145e76edad9ee93243
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2611665
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840662}