[Subresource Filter] Add content settings-based adjustment to WebLayer
This CL incorporates Chrome's adjustment of subresource filter page
activation decisions based on content settings within WebLayer. To do
so, we have the shared SubresourceFilterSafeBrowsingActivationThrottle
directly invoke the logic in question rather than having
ChromeSubresourceFilterClient do so. That change necessitates some
others, also made in this CL:
- WebLayer brings up ProfileInteractionManager.
- SubresourceFilterClient provides an accessor for
  ProfileInteractionManager.
- SubresourceFilterSafeBrowsingActivationThrottle calls the adjustment
  logic via a new Delegate interface rather than doing so via
  SubresourceFilterClient. This interface is implemented by
  ProfileInteractionManager in production while allowing for porting of
  tests that provide custom (or nil) stub logic.

We also add browsertests of this interaction in //weblayer.

Note that this CL does not do any incorporation of the relevant *UI*
for content/site settings of ad blocking in WebLayer; that will be
followup work.

Bug: 1116095
Change-Id: I394e62f12d5209a4302ffa57de9095c861b59613
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2603574
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840681}