[iOS] Omnibox animation on NTP scroll
Animates the NTP's fake omnibox to adjust its width and add a separator
when the NTP is scrolled. Refactors headerSynchronizer and moves it from
ContentSuggestionsCoordinator to NewTabPageViewController. Refactored
synchronizer includes additionalOffset property to compensate for offset
header in the refactored NTP.

The omnibox sticking to the top of the NTP will be added in a follow-up
CL.

Bug: 1114792
Change-Id: Iebb1e7ae9a23772bf03d8cfedb5aed3b0ca16762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2611286
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/master@{#840683}