Support Xbox Series X over usb on Linux and ChromeOS
To support the new XBox Series X controller over usb we simply need to
recognize the device id as an XInput device. The reports are identical
to an XBox One controller so we recognize it as one and it works.

This does not work on macOS because the device seems to need some
initialization that is not provided by macOS.

Bug: 1147578
Change-Id: I494835535d4e3f9dd4bbed657fef2f3e15acbeac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2612116
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Commit-Queue: James Hollyer <jameshollyer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840686}