Android: Handle selection direction change for longpress magnifier
We already have the handle direction information in
OnSelectionBoundsChanged(), so we don't need much modification to it.
Instead, we just need to check if there is an end that equal to the
previous one to decide we should show magnifier for which end.

Magnifier needs to know about which selection boundary was changed,
so we could show Magnifier for that boundary, this change fixes the
case that when the boundary is changed, the Magnifier wasn't moved to
the correct boundary.

Bug: 1157372
Change-Id: I6fd7b4f1b145aef733120225ed94a1be91f20bf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2601111
Reviewed-by: Mohsen Izadi <mohsen@chromium.org>
Commit-Queue: Shimi Zhang <ctzsm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840697}