[GridNG] Computation of static offsets edges and grid area for out of flow items
1. Function ResolveOutOfFlowItemGridLines: The column and row span of
the grid area for the out of flow item is obtained through
GridPositionResolver, and the grid lines of the out of flow item are
computed.

2. Refactor of CacheItemSetIndices: For in flow items, computation of
the set indices remains the same. To compute the indices for out of flow
items we take into account the following:
  a. If both the start and end line of the out of flow item are not
  'auto', then the indices are computed as if they were in flow items.
  b. If either is 'auto', then its value is kNotFound.
  c. If either is attempting to create a new line in the grid, then they
  are treated as 'auto'.

3. Helper method ComputeOffsetAndSize: the computation of the offset and
size of each item was generalized to include the cases where either or
both of the set indices are auto in out of flow items. Additionally,
inline and block edges are computed.

Bug: 1045599
Change-Id: I15b680d1866cdb81b42d626bcd6948ae1b7132ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2567573
Commit-Queue: Ana Sollano Kim <ansollan@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840701}