Migrate `document_size_` to PdfViewPluginBase.
This CL migrates `document_size_` to PdfViewPluginBase and provides
an accessor and a mutator for it, so that it can be used by both
PdfViewWebPlugin and OutOfProcessInstance.

Since `document_size_` is moved to PdfViewPluginBase,
GetDocumentPixelWidth() and GetDocumentPixelHeight() are moved to
PdfViewPluginBase as well, and they can be used identically by both
PdfViewWebPlugin and OutOfProcessInstance.

Bug: 1140629
Change-Id: Ic1a4070f02e99e38adf40a4c15965eb385f874ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2604954
Commit-Queue: Hui Yingst <nigi@chromium.org>
Reviewed-by: K. Moon <kmoon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840716}