[iOS] Return if context is nil in redirect
This CL prevents a crash happening on stable where the app will crash
when receiving a redirect while the |context| is nil.
We know that there are multiple places where the |context| can be nil,
we don't know why. We should treat this in the same way as we treat
other places with no context: by returning early.

Fixed: 990499
Change-Id: Ia80ad38c58d1b52022163990857d98e9745dd5de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2611708
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840726}