[Start] Don't auto delete a tab selected from the tab switcher.
In this CL, we introduce a new flag "keepTab" for a Tab with launch type
FROM_START_SURFACE. When this kind of Tab is selected from the carousel
Tab switcher on the Start surface, or from grid tab switcher, the
"keepTab" is set to be true. This guarantees the tab won't be auto
deleted from the TabModel when the back button is tapped.

More details see doc:
https://docs.google.com/document/d/1mZXgQ0fi_b96vcwy1DkgWunl0FutkFwyt-7PrU-Pgjc/edit#

Bug: 1155139
Change-Id: Ib1e2ce917b0396bbd147ada2ebfc470cffd4718a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2585822
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840737}