Create RFH::IsCrossOriginIsolated and validate it in WebMeasureMemory
This creates a simple accessor in RenderFrameHost to check whether a
frame is cross-origin isolated. It allows features that are gated behind
this policy a way to assert that they are only being invoked when
allowed. Currently not all the information needed to determine this is
available in the browser process so the accessor returns kNotIsolated or
kMaybeIsolated.

This also adds such a check to WebMeasureMemory, which is invoked from
a renderer with OnWebMeasureMemoryRequested. It terminates the
requesting renderer if it is definitely not cross-origin isolated.

Based on http://crrev.com/c/2574730 by chrisha@chromium.org, which added
IsCrossOriginIsolated to RenderProcessHost instead.

BUG=1085129

Change-Id: I2d9d1dd344c85c9358d549fc775282e9736df580
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2578083
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840757}