[Reland] Bento: AltTabMode: Update highlight after switching mode.
This is a reland of http://crrev.com/c/2604907, which was reverted
by http://crrev.com/c/2611090 due to flakiness of the test
ModeSelectionWindowCycleControllerTest.CycleShowsWindowsPerMode.
The issue is that CompleteCycling() activates
a window in another desk, so deskSwitchAnimation takes a screenshot
and sometimes fail if the test ends before it is taken.

Original CL description:

Regardless of where the highlight position is, after switching
alt-tab mode, the highlight will be reset to the first or second
window. This is consistent with highlighting when users first
enter alt-tab mode:
- In general case, highlight the second (next) most recently used
window after the current one.
- In the special case that the most recently used is minimized,
highlight itself (the MRU window), so tabbing into it causes it
to unminimize rather than opening up the next window in the list.

- Track mode switching state in the controller to differentiate
between normal tab pressing and mode switching.
- Add Ash unit tests for both cases.

`ash_unittests --gtest_filter=ModeSelectionWindowCycleControllerTest.SwitchingModeUpdates*WindowHighlight`

Bug: 1157100
Test: Manual test (a video uploaded to crbug) and ash unit tests
Change-Id: I9d3c3ee34e18b40c7d490b990f1a7171d8d8f37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2604907
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Min Chen <minch@chromium.org>
Commit-Queue: Cattalyya Nuengsigkapian <cattalyya@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#840348}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2613266
Cr-Commit-Position: refs/heads/master@{#840767}