[MIX-DL] Make notification announce itself to screen readers.
This CL is a small change to announce the arrival of a MIX-DL alert to
accessibility devices.

We don't have testing infrastructure for accessibility in the downloads
shelf right now, and this is a trivial change, so it's submitted without
tests. I manually verified that this works as expected on Mac OS.

Fixed: 1131084
Change-Id: Ie68a4f6ed5f71a25553deb2a776a7dcd3adf30a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2606669
Reviewed-by: Xing Liu <xingliu@chromium.org>
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840772}