Prepare SiteEngagementService for componentization.
Use PermissionsClient to get HostContentSettingsMap. Only
SiteEngagementService is updated because other files are not going to be
moved to //components. This includes important_sites_util, which is not
needed in //components, and site_engagement_service_unittest, which
depends on Chrome code (such as incognito profiles and
SiteEngagementServiceFactory).

Bug: 1147274
Change-Id: I9506e2482452de1db195ccc0d5652f3def0e2f06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2601451
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840817}