Convert UpdateClientFactory to RepeatingCallback
TestExtensionsBrowserClient allows passing a factory callback that can
be used to create UpdateClients. By nature of being a factory, this
should be a repeating callback. We also pass it by value to indicate a
transfer of ownership.

No change to (the only) call site as it already uses BindRepeating.

Bug: 1152268
Change-Id: Ib07a63a48614fe3bb52b8fc060dd3946f64ec42f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2605874
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840910}