Tests: Make LoggedInUserMixin more user-friendly
LoggedInUserMixin is not as easy to use as it could be. This CL
maximizes the chances that LoggedInUserMixin just works out of the
box, while minimizing the chances of mysterious time outs. The test
developer can optimize for efficiency later by tweaking the options in
LoggedInUserMixin. This CL improves the browser test login tools.

Bug: None
Change-Id: Ic5e597e68ba32e7f29daa4d5331d47298ef3fbcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2598044
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840919}