[printing] Convert PrintHostMsg_RequestPrintPreview to Mojo
This CL converts PrintHostMsg_RequestPrintPreview message to
RequestPrintPreview() in mojom::PrintManagerHost.

Since PrintHostMsg_RequestPrintPreview is a message to
request a print preview, a Preview ID is not generated yet
when the message is requested. So, the converted message is
added to mojom::PrintManagerHost, not mojom::PrintPreviewUI.

This change also updates PrintPreviewDialogControllerBrowserTest
to use TestPrintViewManager instead of
RequestPrintPreviewObserver::OnMessageReceived.

Bug: 1008939
Change-Id: Id5db2862ddb01a91f80cf5534fde71e88c2d2b7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2593847
Reviewed-by: Wei Li <weili@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Julie Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/master@{#840939}